requirements.txt
# Core dependencies
pydantic>=2.4.2
pytest>=7.0.0
pytest-asyncio>=0.20.0
PyYAML>=6.0

# Network and API clients
aiohttp>=3.8.5
requests>=2.28.0

# ZeroMQ for communication
pyzmq>=24.0.0

# Kubernetes for orchestration
kubernetes>=25.0.0

# Data handling
numpy>=1.24.0
pandas>=1.4.0

# Monitoring and metrics
prometheus-client>=0.14.0

# Utilities
click>=8.0.0
tqdm>=4.64.0

# AIPerf requirements
fastapi>=0.104.0
uvicorn>=0.23.2
tiktoken>=0.5.1
python-dotenv>=1.0.0 